+++ FrontEnd/aula-equis/src/router/AppRouter.jsx
@@ -3,7 +3,6 @@ import FormRegister from '../aula-equis/components/FormRegister/FormRegister'
 import { Home } from '../aula-equis/components/Home/Home'
 import FormLogin from '../aula-equis/pages/formLogin/FormLogin'
 import { Landing } from '../aula-equis/pages/landing/Landing'
-import SendEmail from '../aula-equis/pages/sendemail/SendEmail'
 
 export const AppRouter = () => {
   return (
@@ -14,10 +13,6 @@ export const AppRouter = () => {
       <Route path="/login-tutor" element={<FormLogin />} />
       <Route path="/login-escuela" element={<FormLogin />} />
       <Route path="/login-register" element={<FormRegister />} />
-      <Route path="/login-send-email-alumno" element={<SendEmail />} />
-      <Route path="/login-send-email-docente" element={<SendEmail />} />
-      <Route path="/login-send-email-tutor" element={<SendEmail />} />
-      <Route path="/login-send-email-escuela" element={<SendEmail />} />
       <Route path="/form-register" element={<FormRegister />} />
       <Route path="/home" element={<Home />} />
     </Routes>
